【描述】添加释放yuv内存，避免内存泄漏，修改可用容量百分比为float型

+++ storage/storage.h
@@ -39,7 +39,7 @@ S_Result storage_module_init(void);
 
 S_Result storage_module_exit(void);
 
-S_Result storage_sdcard_capacity_info(unsigned int *mbFreedisk, unsigned int *mbTotalsize, unsigned int *percent);
+S_Result storage_sdcard_capacity_info(unsigned int *mbFreedisk, unsigned int *mbTotalsize, float *percent);
 
 
 #endif  /* _CAL_H_ */
